feat: lint and format 2

+++ contracts/src/common/BridgeEncoder.sol
@@ -2,33 +2,31 @@
 pragma solidity ^0.8.13;
 
 import {BridgeMessageType} from "./IBridge.sol";
-import {TransferData, MigrationData} from "./TransferData.sol";
+import {TransferData} from "./TransferData.sol";
 
-/**
- * @dev Library for encoding and decoding bridge messages.
- */
+/// @dev Library for encoding and decoding bridge messages.
 library BridgeEncoder {
+    ////////////////////////////////////////////////////////////////////////
+    // Errors
+    ////////////////////////////////////////////////////////////////////////
+
     /// @dev Error thrown when message type is invalid for ejection
     error InvalidEjectionMessageType();
 
     /// @dev Error thrown when message type is invalid for renewal
     error InvalidRenewalMessageType();
 
-    /**
-     * @dev Encode an ejection message.
-     */
-    function encodeEjection(
-        TransferData memory data
-    ) internal pure returns (bytes memory) {
-        return abi.encode(uint(BridgeMessageType.EJECTION), data);
+    ////////////////////////////////////////////////////////////////////////
+    // Library Functions
+    ////////////////////////////////////////////////////////////////////////
+
+    /// @dev Encode an ejection message.
+    function encodeEjection(TransferData memory data) internal pure returns (bytes memory) {
+        return abi.encode(uint256(BridgeMessageType.EJECTION), data);
     }
 
-    /**
-     * @dev Decode an ejection message.
-     */
-    function decodeEjection(bytes memory message) internal pure returns (
-        TransferData memory data
-    ) {
+    /// @dev Decode an ejection message.
+    function decodeEjection(bytes memory message) internal pure returns (TransferData memory data) {
         uint _messageType;
         (_messageType, data) = abi.decode(message, (uint, TransferData));
         if (_messageType != uint(BridgeMessageType.EJECTION)) {
@@ -36,23 +34,15 @@ library BridgeEncoder {
         }
     }
 
-    /**
-     * @dev Encode a renewal message.
-     */
-    function encodeRenewal(
-        uint256 tokenId,
-        uint64 newExpiry
-    ) internal pure returns (bytes memory) {
+    /// @dev Encode a renewal message.
+    function encodeRenewal(uint256 tokenId, uint64 newExpiry) internal pure returns (bytes memory) {
         return abi.encode(uint(BridgeMessageType.RENEWAL), tokenId, newExpiry);
     }
 
-    /**
-     * @dev Decode a renewal message.
-     */
-    function decodeRenewal(bytes memory message) internal pure returns (
-        uint256 tokenId,
-        uint64 newExpiry
-    ) {
+    /// @dev Decode a renewal message.
+    function decodeRenewal(
+        bytes memory message
+    ) internal pure returns (uint256 tokenId, uint64 newExpiry) {
         uint _messageType;
         (_messageType, tokenId, newExpiry) = abi.decode(message, (uint, uint256, uint64));
         if (_messageType != uint(BridgeMessageType.RENEWAL)) {
@@ -60,9 +50,7 @@ library BridgeEncoder {
         }
     }
 
-    /**
-     * @dev Helper function to get the message type from an encoded message.
-     */
+    /// @dev Helper function to get the message type from an encoded message.
     function getMessageType(bytes memory message) internal pure returns (BridgeMessageType) {
         uint _messageType = abi.decode(message, (uint));
         return BridgeMessageType(_messageType);
