+++ contracts/test/BridgeEncoder.t.sol
@@ -1,24 +1,24 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.13;
 
-import "forge-std/Test.sol";
-import "../src/common/BridgeEncoder.sol";
-import {BridgeMessageType} from "../src/common/IBridge.sol";
-import {TransferData} from "../src/common/TransferData.sol";
-import {NameUtils} from "../src/common/NameUtils.sol";
+// solhint-disable no-console, private-vars-leading-underscore, state-visibility, func-name-mixedcase, namechain/ordering, one-contract-per-file
+
+import {Test} from "forge-std/Test.sol";
+
+import {BridgeEncoder} from "./../src/common/BridgeEncoder.sol";
+import {BridgeMessageType} from "./../src/common/IBridge.sol";
+import {NameUtils} from "./../src/common/NameUtils.sol";
+import {TransferData} from "./../src/common/TransferData.sol";
 
 // Wrapper contract to properly test library errors
 contract BridgeEncoderWrapper {
-    function decodeEjection(bytes memory message) external pure returns (
-        TransferData memory data
-    ) {
+    function decodeEjection(bytes memory message) external pure returns (TransferData memory data) {
         return BridgeEncoder.decodeEjection(message);
     }
 
-    function decodeRenewal(bytes memory message) external pure returns (
-        uint256 tokenId,
-        uint64 newExpiry
-    ) {
+    function decodeRenewal(
+        bytes memory message
+    ) external pure returns (uint256 tokenId, uint64 newExpiry) {
         return BridgeEncoder.decodeRenewal(message);
     }
 }
@@ -42,11 +42,11 @@ contract BridgeEncoderTest is Test {
         });
 
         bytes memory encodedMessage = BridgeEncoder.encodeEjection(transferData);
-        
+
         // Verify the message type is correct
         BridgeMessageType messageType = BridgeEncoder.getMessageType(encodedMessage);
-        assertEq(uint(messageType), uint(BridgeMessageType.EJECTION));
-        
+        assertEq(uint256(messageType), uint256(BridgeMessageType.EJECTION));
+
         // Verify we can decode the message back
         (TransferData memory decodedData) = BridgeEncoder.decodeEjection(encodedMessage);
         assertEq(keccak256(decodedData.dnsEncodedName), keccak256(dnsEncodedName));
@@ -63,13 +63,15 @@ contract BridgeEncoderTest is Test {
         uint64 newExpiry = uint64(block.timestamp + 365 days);
 
         bytes memory encodedMessage = BridgeEncoder.encodeRenewal(tokenId, newExpiry);
-        
+
         // Verify the message type is correct
         BridgeMessageType messageType = BridgeEncoder.getMessageType(encodedMessage);
-        assertEq(uint(messageType), uint(BridgeMessageType.RENEWAL));
-        
+        assertEq(uint256(messageType), uint256(BridgeMessageType.RENEWAL));
+
         // Verify we can decode the message back
-        (uint256 decodedTokenId, uint64 decodedExpiry) = BridgeEncoder.decodeRenewal(encodedMessage);
+        (uint256 decodedTokenId, uint64 decodedExpiry) = BridgeEncoder.decodeRenewal(
+            encodedMessage
+        );
         assertEq(decodedTokenId, tokenId);
         assertEq(decodedExpiry, newExpiry);
     }
@@ -86,10 +88,10 @@ contract BridgeEncoderTest is Test {
             expires: uint64(block.timestamp + 365 days),
             roleBitmap: 0x01
         });
-        
+
         // Manually encode with wrong message type to test custom error
-        bytes memory invalidMessage = abi.encode(uint(BridgeMessageType.RENEWAL), transferData);
-        
+        bytes memory invalidMessage = abi.encode(uint256(BridgeMessageType.RENEWAL), transferData);
+
         // Try to decode it as an ejection message - should revert with custom error
         vm.expectRevert(abi.encodeWithSelector(BridgeEncoder.InvalidEjectionMessageType.selector));
         wrapper.decodeEjection(invalidMessage);
@@ -98,10 +100,14 @@ contract BridgeEncoderTest is Test {
     function testDecodeRenewalInvalidMessageType() public {
         uint256 tokenId = 12345;
         uint64 newExpiry = uint64(block.timestamp + 365 days);
-        
+
         // Manually encode with wrong message type to test custom error
-        bytes memory invalidMessage = abi.encode(uint(BridgeMessageType.EJECTION), tokenId, newExpiry);
-        
+        bytes memory invalidMessage = abi.encode(
+            uint256(BridgeMessageType.EJECTION),
+            tokenId,
+            newExpiry
+        );
+
         // Try to decode it as a renewal message - should revert with custom error
         vm.expectRevert(abi.encodeWithSelector(BridgeEncoder.InvalidRenewalMessageType.selector));
         wrapper.decodeRenewal(invalidMessage);
@@ -117,18 +123,18 @@ contract BridgeEncoderTest is Test {
             expires: uint64(block.timestamp + 365 days),
             roleBitmap: 0x01
         });
-        
+
         // Test ejection message type
         bytes memory ejectionMessage = BridgeEncoder.encodeEjection(transferData);
         BridgeMessageType ejectionType = BridgeEncoder.getMessageType(ejectionMessage);
-        assertEq(uint(ejectionType), uint(BridgeMessageType.EJECTION));
-        
+        assertEq(uint256(ejectionType), uint256(BridgeMessageType.EJECTION));
+
         // Test renewal message type
         uint256 tokenId = 12345;
         uint64 newExpiry = uint64(block.timestamp + 365 days);
         bytes memory renewalMessage = BridgeEncoder.encodeRenewal(tokenId, newExpiry);
         BridgeMessageType renewalType = BridgeEncoder.getMessageType(renewalMessage);
-        assertEq(uint(renewalType), uint(BridgeMessageType.RENEWAL));
+        assertEq(uint256(renewalType), uint256(BridgeMessageType.RENEWAL));
     }
 
     function testEncodingStructure() public view {
@@ -144,10 +150,10 @@ contract BridgeEncoderTest is Test {
         });
 
         bytes memory encodedMessage = BridgeEncoder.encodeEjection(transferData);
-        
+
         // Decode and verify all fields match exactly
         (TransferData memory decodedData) = BridgeEncoder.decodeEjection(encodedMessage);
-        
+
         assertEq(keccak256(decodedData.dnsEncodedName), keccak256(dnsEncodedName));
         assertEq(keccak256(decodedData.dnsEncodedName), keccak256(transferData.dnsEncodedName));
         assertEq(decodedData.owner, transferData.owner);
@@ -156,4 +162,4 @@ contract BridgeEncoderTest is Test {
         assertEq(decodedData.expires, transferData.expires);
         assertEq(decodedData.roleBitmap, transferData.roleBitmap);
     }
-} 
+}
